+++ ErsatzTV.Application/Graphics/Commands/RefreshGraphicsElementsHandler.cs
@@ -1,6 +1,7 @@
 using ErsatzTV.Core;
 using ErsatzTV.Core.Domain;
 using ErsatzTV.Core.Interfaces.Metadata;
+using ErsatzTV.Core.Interfaces.Streaming;
 using ErsatzTV.Infrastructure.Data;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
@@ -10,6 +11,7 @@ namespace ErsatzTV.Application.Graphics;
 public class RefreshGraphicsElementsHandler(
     IDbContextFactory<TvContext> dbContextFactory,
     ILocalFileSystem localFileSystem,
+    IGraphicsElementLoader graphicsElementLoader,
     ILogger<RefreshGraphicsElementsHandler> logger)
     : IRequestHandler<RefreshGraphicsElements>
 {
@@ -21,7 +23,11 @@ public class RefreshGraphicsElementsHandler(
         List<GraphicsElement> allExisting = await dbContext.GraphicsElements
             .ToListAsync(cancellationToken);
 
-        foreach (GraphicsElement existing in allExisting.Where(e => !localFileSystem.FileExists(e.Path)))
+        var missing = allExisting
+            .Where(e => !localFileSystem.FileExists(e.Path) || Path.GetExtension(e.Path) != ".yml")
+            .ToList();
+
+        foreach (GraphicsElement existing in missing)
         {
             logger.LogWarning(
                 "Removing graphics element that references non-existing file {File}",
@@ -30,8 +36,13 @@ public class RefreshGraphicsElementsHandler(
             dbContext.GraphicsElements.Remove(existing);
         }
 
+        foreach (GraphicsElement existing in allExisting.Except(missing))
+        {
+            await TryRefreshName(existing, cancellationToken);
+        }
+
         // add new text elements
-        var newTextPaths = localFileSystem.ListFiles(FileSystemLayout.GraphicsElementsTextTemplatesFolder)
+        var newTextPaths = localFileSystem.ListFiles(FileSystemLayout.GraphicsElementsTextTemplatesFolder, "*.yml")
             .Where(f => allExisting.All(e => e.Path != f))
             .ToList();
 
@@ -45,11 +56,13 @@ public class RefreshGraphicsElementsHandler(
                 Kind = GraphicsElementKind.Text
             };
 
+            await TryRefreshName(graphicsElement, cancellationToken);
+
             await dbContext.AddAsync(graphicsElement, cancellationToken);
         }
 
         // add new image elements
-        var newImagePaths = localFileSystem.ListFiles(FileSystemLayout.GraphicsElementsImageTemplatesFolder)
+        var newImagePaths = localFileSystem.ListFiles(FileSystemLayout.GraphicsElementsImageTemplatesFolder, "*.yml")
             .Where(f => allExisting.All(e => e.Path != f))
             .ToList();
 
@@ -63,11 +76,13 @@ public class RefreshGraphicsElementsHandler(
                 Kind = GraphicsElementKind.Image
             };
 
+            await TryRefreshName(graphicsElement, cancellationToken);
+
             await dbContext.AddAsync(graphicsElement, cancellationToken);
         }
 
         // add new motion elements
-        var newMotionPaths = localFileSystem.ListFiles(FileSystemLayout.GraphicsElementsMotionTemplatesFolder)
+        var newMotionPaths = localFileSystem.ListFiles(FileSystemLayout.GraphicsElementsMotionTemplatesFolder, "*.yml")
             .Where(f => allExisting.All(e => e.Path != f))
             .ToList();
 
@@ -81,11 +96,13 @@ public class RefreshGraphicsElementsHandler(
                 Kind = GraphicsElementKind.Motion
             };
 
+            await TryRefreshName(graphicsElement, cancellationToken);
+
             await dbContext.AddAsync(graphicsElement, cancellationToken);
         }
 
         // add new subtitle elements
-        var newSubtitlePaths = localFileSystem.ListFiles(FileSystemLayout.GraphicsElementsSubtitleTemplatesFolder)
+        var newSubtitlePaths = localFileSystem.ListFiles(FileSystemLayout.GraphicsElementsSubtitleTemplatesFolder, "*.yml")
             .Where(f => allExisting.All(e => e.Path != f))
             .ToList();
 
@@ -99,9 +116,21 @@ public class RefreshGraphicsElementsHandler(
                 Kind = GraphicsElementKind.Subtitle
             };
 
+            await TryRefreshName(graphicsElement, cancellationToken);
+
             await dbContext.AddAsync(graphicsElement, cancellationToken);
         }
 
         await dbContext.SaveChangesAsync(cancellationToken);
+    }
+
+    private async Task TryRefreshName(GraphicsElement graphicsElement, CancellationToken cancellationToken)
+    {
+        graphicsElement.Name = null;
+        Option<string> maybeName = await graphicsElementLoader.TryLoadName(graphicsElement.Path, cancellationToken);
+        foreach (string name in maybeName)
+        {
+            graphicsElement.Name = name;
+        }
     }
 }