+++ ErsatzTV.Core/Domain/GraphicsElement.cs
@@ -6,6 +6,7 @@ public class GraphicsElement
 {
     public int Id { get; set; }
     public string Path { get; set; }
+    public string Name { get; set; }
     public GraphicsElementKind Kind { get; set; }
     public List<PlayoutItem> PlayoutItems { get; set; }
     public List<PlayoutItemGraphicsElement> PlayoutItemGraphicsElements { get; set; }